Create sparse matrix of constraints in sorted order
If matrix is not sorted, it will get sorted inplace eventually, e.g., when calling min or max methods.
With scipy 0.18, this is a problem, because matrix.T.max() (called by cvxpy) would sort the transposed matrix inplace,
which does affect the data in original matrix as well. Consequently, the values of the original matrix will
be wrong.

https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.csr_matrix.sort_indices.html#scipy.sparse.csr_matrix.sort_indices