CI: Add support for appveyor builds
Install dependencies for Python 3.4:
cvxopt is not available for python >=3.5
(cvxopt/cvxopt#94)

Install numpy, pandas, and scipy via conda.
Install the remainder via pip, because
conda package of cvxpy and scs is not available
for Windows, yet.

Added hack to activate Miniconda for Python 3.4.

Run nosetests:
- Skip slow tests.
- Skip test_breast_cancer_rbf_cvxopt, because
performace is slightly different in Windows and
has not been excluded via -a "!slow".