Update code to compute set of comparable pairs
Changes the tied_time return value of concordance_index_censored.
It is now the number of comparable pairs that share the same time
and that are used in computing the concordance index.
Previously, the number was not directly related to the concordance
index itself.
We consider a pair with the same time as comparable, if one is an
event and the other is censored.