Fix estimation of censoring distribution for tied times with events
Adds a reverse argument to kaplan_meier_estimator

When estimating the censoring distribution, we now consider events
to occur before censoring. For tied time points with an event, those
with an event are not considered at risk anymore and subtracted from
the demoniator of the Kaplan-Meier estimator.

The change affects all functions relying on inverse probability
of censoring weights, namely:
- sksurv.nonparametric.CensoringDistributionEstimator
- sksurv.nonparametric.ipc_weights
- sksurv.linear_model.IPCRidge
- sksurv.metrics.cumulative_dynamic_auc
- sksurv.metrics.concordance_index_ipcw