Improvement of calculating cumulative dynamic auc/roc
Calculation of AUC is faster due to use of numpy routines instead of for loops.
In addition this generalizes the time depending ROC/AUC to time-varying estimates
(e.g. predicted survival function from RSF).
Instead of (n_samples,) shaped estimate (n_samples, n_times,) are given, where the
estimate for each timepoint in times, the AUC/ROC should be calculated for should
be given as input along axis=1.
By specifiying ret_roc=True, the tpr and fpr for each timepoint in times is returned
and can be used to plot time-depedent roc curves.