Fix column.standardize uses different std for numpy arrays
numpy has default ddof=0, whereas pandas has ddof=1

Do not modify passed DataFrame or array in-place.
This fixes a problem in tests, where test using DataFrame was
executed first and modified the underlying numpy array,
which made the subsequent test based on numpy array pass.
However, calling standardize on an object numpy array leaves
it unchanged, which was not reflected in the test.