FIX Wrong baseline functions in CoxnetSurvivalAnalysis
Features are centered before fitting and afterwards the
baseline survival function and CHF are estimated from
the predicted risk scores, using the centered data.
However, when original data is not centered, predictions
of survival and CHF will be wrong, because they expect
centered data.

We can ensure predictions are always on the centered scale by
computing an offset term (foreach alpha) that is subtracted
from the standard linear predictor X @ coef.

Fixes #139