Add confidence intervals to the cumulative incidence estimator for competing risks.
Tests are included.
There are some minimal differences in the results of the confidence
intervals with respect to the implementation in
the R package cmprsk, so we add some tolerance to the assertions.
Apart from the Aalen estimator used in cmprsk we include two other
estimators, but these are not tested so thoroughly, because we couldn't
find data sources.