Load expected data from external file
If var_type is one of "Aalen", "Dinse_Approx" the true
values are based on the proc lifetest statement in SAS:
https://support.sas.com/documentation/onlinedoc/stat/ex_code/143/liftcrsk.html

Because an external implementation of the exact Dinse variance
has not been found, the true CI values are taken from this internal code.
Thus, this serves as a regression test.